+++ src/components/DivNote.tsx
@@ -1,7 +1,7 @@
 import { makeStyles, Theme } from '@material-ui/core'
 import React, { FunctionComponent, useEffect, useState } from 'react'
-import { Note } from '../firestore/types/note'
-import { watchNote } from '../firestore/watchNote'
+import { Note } from '../shared/firestore/types/note'
+import { watchNote } from '../shared/firestore/watchNote'
 import { UpdateNoteData } from '../shared/functions/types/updateNoteData'
 import { updateNote } from '../shared/functions/updateNote'
 import DivNoteEditor from './DivNoteEditor'
