--- src/firestore/constants/collection.ts
@@ -1,1 +1,0 @@
-export const NOTES = 'notes'
