+++ src/shared/firebase/initializeApp.ts
