+++ src/shared/firestore/deleteNote.ts
