+++ src/shared/functions/createNote.ts
@@ -1,6 +1,6 @@
 import { app } from 'firebase/app'
 import { httpsCallable } from 'rxfire/functions'
-import { ASIA_NORTHEAST1 } from '../constants/region'
+import { ASIA_NORTHEAST1 } from './constants/region'
 import { CreateNoteData } from './types/createNoteData'
 import { CreateNoteResult } from './types/createNoteResult'
