+++ src/shared/functions/updateNote.ts
@@ -1,6 +1,6 @@
 import { app } from 'firebase/app'
 import { httpsCallable } from 'rxfire/functions'
-import { ASIA_NORTHEAST1 } from '../constants/region'
+import { ASIA_NORTHEAST1 } from './constants/region'
 import { UpdateNoteData } from './types/updateNoteData'
 import { UpdateNoteResult } from './types/updateNoteResult'
