+++ src/shared/helpers/createMarkup.ts
@@ -1,6 +1,7 @@
 import marked from 'marked'
 
 marked.setOptions({ breaks: true })
+
 export const createMarkup = (text: string) => {
   return { __html: marked(text) }
 }
