+++ src/shared/operators/mapEmpty.ts
