+++ src/shared/operators/mapQuerySnapToDoc.ts
