+++ src/shared/operators/mapQuerySnapToDocs.ts
