+++ src/shared/operators/mapSnapToDoc.ts
